# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Kreatorboard API - Creator Operating System
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /profile:
    get:
      operationId: getMyProfile
      tags: [profile]
      summary: Get my creator profile
      responses:
        "200":
          description: Creator profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatorProfile"
        "404":
          description: Profile not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateMyProfile
      tags: [profile]
      summary: Update my creator profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatorProfile"
  /profile/onboarding:
    post:
      operationId: completeOnboarding
      tags: [profile]
      summary: Complete onboarding step
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OnboardingStepRequest"
      responses:
        "200":
          description: Onboarding updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OnboardingResponse"
...
  /profile/public/{handle}:
    get:
      operationId: getPublicProfile
      tags: [profile]
      summary: Get public creator profile by handle
      parameters:
        - name: handle
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Public profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicCreatorProfile"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /profile/services:
    get:
      operationId: getMyServices
      tags: [profile]
      summary: Get my services list
      responses:
        "200":
          description: Services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CreatorService"
    post:
      operationId: createService
      tags: [profile]
      summary: Create a service
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateServiceRequest"
      responses:
        "201":
          description: Created service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatorService"
  /profile/services/{id}:
    patch:
      operationId: updateService
      tags: [profile]
      summary: Update a service
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateServiceRequest"
      responses:
        "200":
          description: Updated service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatorService"
    delete:
      operationId: deleteService
      tags: [profile]
      summary: Delete a service
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /profile/portfolio:
    get:
      operationId: getMyPortfolio
      tags: [profile]
      summary: Get my portfolio items
      responses:
        "200":
          description: Portfolio items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PortfolioItem"
    post:
      operationId: createPortfolioItem
      tags: [profile]
      summary: Create portfolio item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePortfolioItemRequest"
      responses:
        "201":
          description: Created item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortfolioItem"
...
  /payments/requests:
    get:
      operationId: getPaymentRequests
      tags: [payments]
      summary: Get all payment requests
      responses:
        "200":
          description: Payment requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PaymentRequest"
    post:
      operationId: createPaymentRequest
      tags: [payments]
      summary: Create a payment request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePaymentRequestInput"
      responses:
        "201":
          description: Created payment request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentRequest"
  /payments/requests/{id}:
    get:
      operationId: getPaymentRequest
      tags: [payments]
      summary: Get a payment request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Payment request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentRequest"
    patch:
      operationId: updatePaymentRequest
      tags: [payments]
      summary: Update payment request status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePaymentRequestInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentRequest"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    CreatorProfile:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        handle:
          type: string
        displayName:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        bannerUrl:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        timezone:
          type: string
          nullable: true
        website:
          type: string
          nullable: true
        niches:
          type: array
          items:
            type: string
        ecosystems:
          type: array
          items:
            type: string
        languages:
          type: array
          items:
            type: string
        isPublic:
          type: boolean
        profileCompleteness:
          type: integer
        calComLink:
          type: string
          nullable: true
        linktreeUrl:
          type: string
          nullable: true
        twitterHandle:
          type: string
          nullable: true
        discordHandle:
          type: string
          nullable: true
        telegramHandle:
          type: string
          nullable: true
        openToCollaboration:
          type: boolean
        mediaKitUrl:
          type: string
          nullable: true
        publicPageUrl:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - handle
        - isPublic
        - openToCollaboration
        - profileCompleteness
        - niches
        - ecosystems
        - languages
        - createdAt
        - updatedAt
    PublicCreatorProfile:
      type: object
      properties:
        handle:
          type: string
        displayName:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        bannerUrl:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        niches:
          type: array
          items:
            type: string
        ecosystems:
          type: array
          items:
            type: string
        languages:
          type: array
          items:
            type: string
        website:
          type: string
          nullable: true
        calComLink:
          type: string
          nullable: true
        socialLinks:
          type: array
          items:
            $ref: "#/components/schemas/SocialLink"
        services:
          type: array
          items:
            $ref: "#/components/schemas/CreatorService"
        portfolio:
          type: array
          items:
            $ref: "#/components/schemas/PortfolioItem"
        openToCollaboration:
          type: boolean
        connectedWalletAddress:
          type: string
          nullable: true
      required:
        - handle
        - niches
        - ecosystems
        - languages
        - socialLinks
        - services
        - portfolio
        - openToCollaboration
    SocialLink:
      type: object
      properties:
        platform:
          type: string
        url:
          type: string
        handle:
          type: string
          nullable: true
      required:
        - platform
        - url
    UpdateProfileRequest:
      type: object
      properties:
        displayName:
          type: string
        bio:
          type: string
        location:
          type: string
        timezone:
          type: string
        website:
          type: string
        niches:
          type: array
          items:
            type: string
        ecosystems:
          type: array
          items:
            type: string
        languages:
          type: array
          items:
            type: string
        isPublic:
          type: boolean
        calComLink:
          type: string
        linktreeUrl:
          type: string
        openToCollaboration:
          type: boolean
    OnboardingStepRequest:
      type: object
      properties:
        step:
          type: integer
        data:
          type: object
          additionalProperties: true
      required:
        - step
    OnboardingResponse:
      type: object
      properties:
        currentStep:
          type: integer
        complete:
          type: boolean
        nextAction:
          type: string
          nullable: true
      required:
        - currentStep
        - complete
    CreatorService:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
          nullable: true
        price:
          type: string
          nullable: true
        currency:
          type: string
          nullable: true
        deliveryTime:
          type: string
          nullable: true
        isActive:
          type: boolean
        category:
          type: string
          nullable: true
      required:
        - id
        - title
        - isActive
    CreateServiceRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        price:
          type: string
        currency:
          type: string
        deliveryTime:
          type: string
        category:
          type: string
        isActive:
          type: boolean
      required:
        - title
    PortfolioItem:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        url:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
        platform:
          type: string
          nullable: true
        featured:
          type: boolean
        sortOrder:
          type: integer
      required:
        - id
        - title
        - featured
        - sortOrder
    CreatePortfolioItemRequest:
      type: object
      properties:
        title:
          type: string
        url:
          type: string
        description:
          type: string
        imageUrl:
          type: string
        platform:
          type: string
        featured:
          type: boolean
      required:
        - title
...
    PaymentRequest:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        amount:
          type: string
        currency:
          type: string
        toAddress:
          type: string
        fromAddress:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, paid, cancelled, expired]
        paymentLink:
          type: string
          nullable: true
        serviceId:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        paidAt:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - title
        - amount
        - currency
        - toAddress
        - status
        - createdAt
    CreatePaymentRequestInput:
      type: object
      properties:
        title:
          type: string
        amount:
          type: string
        currency:
          type: string
        walletId:
          type: string
        serviceId:
          type: string
        notes:
          type: string
      required:
        - title
        - amount
        - currency
        - walletId
    UpdatePaymentRequestInput:
      type: object
      properties:
        status:
          type: string
          enum: [pending, paid, cancelled, expired]
        fromAddress:
          type: string
        notes:
          type: string
...